refactoring replace sight details screen
