+++ lib/main.dart
@@ -3,7 +3,7 @@ import 'package:places/constants/app_strings.dart';
 import 'package:places/mocks.dart';
 import 'package:places/res/themes.dart';
 import 'package:intl/date_symbol_data_local.dart';
-import 'package:places/ui/screen/sight_details_screen.dart';
+import 'package:places/ui/screen/sight_details_screen/sight_details_screen.dart';
 
 void main() {
   initializeDateFormatting();
