+++ lib/ui/screen/sight_details_screen/sight_details_screen.dart
